Refactored accountsClearSelection feature to `sauce`.
