+++ sauce/features/accounts/clear-selection/settings.js
@@ -1,0 +1,8 @@
+module.exports = {
+  name: 'ClearSelection',
+  type: 'checkbox',
+  default: false,
+  section: 'accounts',
+  title: 'Clear Selection',
+  description: 'Adds an option to the transaction edit drop-down menu to clear the current selection.'
+};
