--- source/common/res/features/accounts-clear-selection/settings.json
@@ -1,13 +1,0 @@
-{
-         "name": "accountsClearSelection",
-         "type": "checkbox",
-      "default": false,
-      "section": "accounts",
-        "title": "Clear Selection",
-  "description": "Adds an option to the transaction edit drop-down menu to clear the current selection.",
-      "actions": {
-                    "true": [
-                      "injectScript", "main.js"
-                    ]
-                 }
-}
